move template.py into actual package
